Permitir decimales en los inputs

+++ database/migrations/2023_02_13_0_create_viaje_inicial.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('salida');
             $table->date('dia2')->nullable();
             $table->string('llegada');
-            $table->integer('TN');
+            $table->decimal('TN');
             $table->timestamps();
         });
     }
